refactor day 06 to modern style
This one was a bear. It wasn't bad in itself, but it turns out that
`parse-display` has a bug which meant that it just cannot handle this input:
<https://github.com/frozenlib/parse-display/issues/9>. Working around
that issue meant breaking out the big guns (LALRPOP), which while fun,
does take some extra time.

That said, I'm pretty happy about `trait ManipulateLight`; it feels
like a _much_ more elegant solution to the problem of abstracting
over the fundamental type of the map than what I'd previously had.
I would not have been likely to come up with it if I hadn't known
from the start that I'd need varying map tile types, but since I did,
it was pretty nice to be able to just write the right thing.

The trait bounds in `Command::apply` are definitely past what I could
do in 2015, though!

Also, this reflects better on 2019 me than current me, but I'm pretty
happy with how well the Map and Point classes applied to this problem.

I realized partway into implementation that overflow was a possibility
given that there are 300 instructions and only 256 values in a u8,
so it was nice to see that everything worked appropriately.

This is the first exercise of 2015 that took substantially longer to
run in debug than release mode. At a guess, this has to do with the
parser, but I'm not willing to invest the effort to profile it right now.