refactor day 07 to modern style
This was fun: there was _quite a lot_ of old, manual parsing code,
none of which is now required. LALRPOP was absolutely the right choice
given the parsing requirements today.

I think that last time I did this, I might have gone through the
effort to properly topo-sort the graph, then resolve it in appropriate
sequence. This time, I just looped over the set of wires, calculating
everything for which I had the appropriate information, and then
repeating the process until there were no more uncomputed wires.
It's much simpler, and fast enough.

That may as well be the motto of these writes: "It's much simpler,
and fast enough."