WIP: add rust implementation of day 08
This one is kind of weird: part 1 is straight-up wrong, despite
passing the tests. Not sure what's up there.

The actual answer for my input is 1371. The answer this version produces
is 1310, which means that some characters are not making it into
`unescape`'s output, which should.

Also, when I look at the python implementation, it looks significantly
rustier than the actual rust code I wrote. Not sure what's up with
that weird state machine in retrospect. Will likely end up rewriting
to use an explicit enum-based state machine later.