Fix day 08 implementation
This one was a head-scratcher. The rust code was persistently and
reliably doing what looked like the right thing, but also very
consistently reporting an answer too short: 1310, instead of 1371.
The simple explanation for a shortfall like that is that we're
sometimes writing too many characters to the output.

Wrote a whole roundtrip CLI flag to try to identify the cases where
that happened, but that just made it worse: in all cases, an E -> U
roundtrip worked perfectly, and a U -> E roundtrip worked losslessly
for all inputs not containing hexadecimal; that exception was just
because the encode step didn't attempt to encode hex.

Compared to the python implementation; it was _very similar_. Felt
like I was crazy. This is the high-octane debugging experience!

Finally realized that rust strings are utf-8 encoded internally,
and the number of bytes (as reported by `.len()`) may not match
the number of chars (`.chars().count()`). Tried that; got the
right answer instantly.

Chagrin, but then again: I win.