refactor day 10 to modern style
Not entirely happy with the algorithm here, but there's not any
obvious improvement I could make. Feels like I should be operating
on a vector of numbers, but that's complicated by the fact that
we then immediately append the characters of the string representation
of the number.

Pretty boring refactor, unfortunately.