refactor day 11 to modern style
This ended up being a complete rewrite, because it's much easier
to treat this problem as mutation of a vector of bytes than as
a problem of generating new strings. Probably faster, too!

Dropped several microoptimizations, but honestly, this runs pretty
fast as-is. I suspect that what we lose in raw number of passwords
checked, we gain in cache and pipelining. Either way, I'd want to
create some good microbenchmarks before actually attempting to
re-enable any of them.