refactor day 15 to modern style
Now that I've had a chance to examine the problem while properly awake,
I'm pretty satisfied with the work I've done rewriting day 15. As a
data-modeling exercise, it was good to separate out the quantities as
the minimal amount of data that must be passed around to uniquely
identify a particular recipe, given that we can preserve the list of
ingredients and their properties separately.

I suspect that in a few months, we'll be able to improve on this
implementation with the stabilization of min_const_generics: instead
of heap-allocating all those little vectors of ingredients, those could
just be heap-allocated arrays. Won't get around to that for a while,
though.

I'm also particularly pleased with the `neighbors_of` function, which
replaces a _much_ more complicated implementation.

Here's a single run of a simple, dumb benchmark. Take it for all that's worth.
(Some extraneous output lines were trimmed.)

$ git checkout 63f31d32d140d1406e43691b8c70667ab8914549 &&
> cargo build --release -p day15 &&
> time target/release/day15 --part2 --no-part1 &&
> git checkout master &&
> cargo build --release -p day15 &&
> time target/release/day15 --part2 --no-part1
HEAD is now at 63f31d3 partially refactor day 15 to modern style
    Finished release [optimized] target(s) in 3.08s
best recipe goodness (constrained to 500 calories): 1766400

real    0m4.108s
user    0m4.088s
sys     0m0.020s
Switched to branch 'master'
    Finished release [optimized] target(s) in 1.98s
best recipe goodness (constrained to 500 calories): 1766400

real    0m0.181s
user    0m0.162s
sys     0m0.010s

It's not definitive, but it sure looks like a ~40x speed improvement.
I'll take it!