rewrite loadout_generator
It turns out that I'm _much_ more comfortable with iterators than
I used to be when I wrote the first draft of this code; that's good,
because this implementation is substantially shorter and simpler.
Can't say that it's more correct given that it still produces the
same answer, but it's a demonstrably better factoring.