start rewriting day24
The hard part of this day's exercise is efficiently generating the
subsets. Permutation-based approaches are insufficient: there are
just too many possible permutations; generating those and then
filtering for sum-suitability just isn't viable.

The approach in `SummedSubsets` is viable, but is currently
implemented in an inefficient manner. In short: a recursive iterator
can generate subsets by greedily seeking through the list, then
recursing to generate the remaining items in the list to form
the appropriate subset.

It still needs rewriting, but it's viable.