BROKEN start working on a re-refactor of BoundedPermutationGenerator
Moved the WIP stuff into a new (temporary) sub-crate because that
makes it much easier to isolate the appropriate warnings/errors.

Fundamentally the issue here is that rustc and I disagree about what's
safe to do with references. I'm convinced that it's safe to implement
this recursive slice-modification algorithm, rustc disagrees for now.

It hasn't yet convinced me that I'm wrong; I'm still hopeful that
with appropriate syntax, I'll be able to convince it that what I'm
doing here is safe and appropriate.